jupyter: refactor the kernel selection code to not pass state down for no reason -- makes the code cleaner, smaller, and easier to reuse and read.
